chart1: Refactor, split out messagebox.cpp/.h.
Chip away some lines from the too big chart1.cpp inte a separate library.
Some files can now just use the small gui_lib dependency instead of the
chart1, cleaning up the dependency graph.
