+++ include/ais.h
@@ -46,6 +46,7 @@
 
 #include "dychart.h"
 #include "chart1.h"
+#include "gui_lib.h"
 #include "navutil.h"
 #include "OCPN_Sound.h"
 #include "AIS_Bitstring.h"
