+++ src/S57QueryDialog.cpp
@@ -28,7 +28,7 @@
 #include "S57QueryDialog.h"
 //#include "chcanv.h"
 #include "navutil.h"
-#include "chart1.h"
+#include "gui_lib.h"
 
 extern ColorScheme global_color_scheme;
 extern S57QueryDialog *g_pObjectQueryDialog;
